A unification of subroutine and block frame creation.
Now both O_subframe() and O_blkframe() opcodes call the
new function execute_in_frame(). The only difference is that
the creation of the l array is forced if it's NULL or if a
subroutine frame is wanted. This change also makes possible
to use arguments when executing code inside a block frame.

This change should *NOT* be disruptive.